requirements.txt
bleach==5.0.0
boto3==1.23.10
Django==4.0.9
django-anymail==8.6
django-cors-headers==3.11.0
django-debug-toolbar==3.7.0
django-hijack==3.1.4
django-multiselectfield==0.1.12
django-quill-editor==0.1.40
django-storages==1.12.3
django-versatileimagefield==2.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
drf-yasg==1.21.4
factory-boy==3.2.1
Faker~=13.11.1
freezegun==1.2.2
getconf==1.11.0
gunicorn==20.1.0
ipython==8.0.1
linkpreview==0.2.2
Pillow==9.4.0
rollbar==0.16.2
telescoop-auth==0.0.8
telescoop-backup==0.1.5
tqdm~=4.64.0
